# Feature/kazuki/optimize bulk query