--- a/app/controllers/api/v1/expenditure_logs_controller.rb
+++ b/app/controllers/api/v1/expenditure_logs_controller.rb
@@ -32,18 +32,20 @@ module Api
       end
 
       def bulk_delete
-        # OPTIMIZE: 可能ならSQL delete処理にしたい。
         begin
-          ActiveRecord::Base.transaction do
-            expenditure_logs = @current_user.expenditure_logs
+          existing_log_ids = @current_user.expenditure_logs.ids
 
-            expenditure_log_ids.each do |id|
-              expenditure_logs.find(id)
-                              .try(:destroy!)
-            end
+          return response_bad_request unless expenditure_log_ids.all? { |id| existing_log_ids.include?(id) }
 
-            response_success(:expenditure_logs, :delete)
-          end
+          query = "DELETE FROM `expenditure_logs` WHERE `expenditure_logs`.`id` IN (#{expenditure_log_ids})"
+          query.delete!('[]')
+          ActiveRecord::Base.connection.execute(query)
+
+          query = "DELETE FROM `tag_relations` WHERE `tag_relations`.`expenditure_log_id` IN (#{expenditure_log_ids})"
+          query.delete!('[]')
+          ActiveRecord::Base.connection.execute(query)
+
+          response_success(:expenditure_logs, :delete)
         rescue => exception
           response_internal_server_error
         end