--- a/app/controllers/api/v1/income_logs_controller.rb
+++ b/app/controllers/api/v1/income_logs_controller.rb
@@ -35,15 +35,18 @@ module Api
         begin
           existing_log_ids = @current_user.income_logs.ids
 
-          if income_log_ids.all? { |id| existing_log_ids.include?(id) }
-            query = "DELETE FROM `income_logs` WHERE `income_logs`.`id` IN (#{income_log_ids})"
-            query.delete!('[]')
-            ActiveRecord::Base.connection.execute(query)
+          # パラメータの中身が全て
+          # existing_log_idsに一致しないならば処理中止
+          return response_bad_request unless income_log_ids.all? { |id| existing_log_ids.include?(id) }
 
-            query = "DELETE FROM `tag_relations` WHERE `tag_relations`.`income_log_id` IN (#{income_log_ids})"
-            query.delete!('[]')
-            ActiveRecord::Base.connection.execute(query)
-          end
+          query = "DELETE FROM `income_logs` WHERE `income_logs`.`id` IN (#{income_log_ids})"
+          query.delete!('[]')
+          ActiveRecord::Base.connection.execute(query)
+
+          # ログに関連するtag relationを一括削除
+          query = "DELETE FROM `tag_relations` WHERE `tag_relations`.`income_log_id` IN (#{income_log_ids})"
+          query.delete!('[]')
+          ActiveRecord::Base.connection.execute(query)
 
           response_success(:income_logs, :delete)
         rescue => exception