--- a/app/models/tag.rb
+++ b/app/models/tag.rb
@@ -2,7 +2,7 @@ class Tag < ApplicationRecord
   # create update時
   # before_save { self.color = '#' + color } // TODO: 正規表現 ?
 
-  has_many :tag_relations, dependent: :destroy
+  has_many :tag_relations, dependent: :delete_all
   has_many :expenditure_logs, through: :tag_relations
   has_many :income_logs, through: :tag_relations
   belongs_to :user