## Changes
一括削除処理のクエリを改善

```
Unpermitted parameter: :income_log
Unpermitted parameter: :income_log
   (37.6ms)  DELETE FROM `income_logs` WHERE `income_logs`.`id` IN (90, 91, 94, 95, 96)
  ↳ app/controllers/api/v1/income_logs_controller.rb:44:in `bulk_delete'
Unpermitted parameter: :income_log
   (4.6ms)  DELETE FROM `tag_relations` WHERE `tag_relations`.`income_log_id` IN (90, 91, 94, 95, 96)
  ↳ app/controllers/api/v1/income_logs_controller.rb:49:in `bulk_delete'
Completed 200 OK in 55ms (Views: 0.3ms | ActiveRecord: 43.0ms | Allocations: 3117)
```

## Related links

## Test result

* [x] featureテストOK
* [x] log一括削除時、タグ削除時クエリ速度改善(sql delete処理)
* [x] sqlインジェクションされない？根拠は？
>クエリ実行前にパラメータ(数字の配列) が `全て` リクエストを実行したユーザーが保持するログのidと一致するかどうかを確認しているため。

例) expenditure logを一括削除
```ruby
existing_log_ids = @current_user.expenditure_logs.ids
# 1 ユーザーに紐づくexpenditure logのidを全て取得
return response_bad_request unless expenditure_log_ids.all? { |id| existing_log_ids.include?(id) }
# 2 expenditure_log_ids(整数の配列)の中身が全て1の配列に含まれていないならクエリ実行中止
```

## Others
テーブルのレンダリングが遅い